# backend/openapi/openapi-draft.yaml
openapi: "3.0.3"
info:
  title: NewHaven API
  version: "0.1.0"
  description: Backend API for NewHaven browser-based economy simulation game
servers:
  - url: http://localhost:8088
paths:
...
  /api/v3/companies/me/buildings/:
    get:
      tags: [Buildings]
      operationId: listMyBuildings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Building list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingListResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v2/buildings/buy/:
    post:
      tags: [Buildings]
      summary: Buy a building
      operationId: buyBuilding
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuyBuildingRequest'
      responses:
        "200":
          description: Building bought
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuyBuildingResponse'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
  /api/v2/buildings/place/:
    post:
      tags: [Buildings]
      summary: Place a building on the map
      operationId: placeBuilding
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceBuildingRequest'
      responses:
        "200":
          description: Building placed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlaceBuildingResponse'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
  /api/v2/buildings/move/:
    post:
      tags: [Buildings]
      summary: Move a building on the map
      operationId: moveBuilding
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveBuildingRequest'
      responses:
        "200":
          description: Building moved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MoveBuildingResponse'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
  /api/v2/buildings/demolish/:
    post:
      tags: [Buildings]
      summary: Demolish a building
      operationId: demolishBuilding
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DemolishBuildingRequest'
      responses:
        "200":
          description: Building demolished
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemolishBuildingResponse'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
  /api/v2/buildings/{buildingId}/stock/:
    post:
      tags: [Buildings]
      summary: Stock items into a retail building's shelf
      operationId: stockShelf
      security:
        - BearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockShelfRequest'
      responses:
        "200":
          description: Shelf stocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShelfActionResponse'
        "400":
          description: Bad request
        "401":
          description: Unauthorized
  /api/v2/buildings/{buildingId}/unstock/:
    post:
      tags: [Buildings]
      summary: Unstock items from a retail building's shelf back to warehouse
      operationId: unstockShelf
      security:
        - BearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnstockShelfRequest'
      responses:
        "200":
          description: Shelf unstocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShelfActionResponse'
        "400":
          description: Bad request
        "401":
          description: Unauthorized
  /api/v2/buildings/{buildingId}/shelf-price/:
    post:
      tags: [Buildings]
      summary: Set price for a shelf item
      operationId: setShelfPrice
      security:
        - BearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetShelfPriceRequest'
      responses:
        "200":
          description: Shelf price updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShelfActionResponse'
        "400":
          description: Bad request
        "401":
          description: Unauthorized
  /api/v1/buildings/{buildingId}/upgrade/:
    post:
      tags: [Buildings]
      summary: Upgrade a building
      operationId: upgradeBuilding
      security:
        - BearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Building upgraded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpgradeBuildingResponse'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
...
  /api/bonds/:
    get:
      tags: [Bonds]
      summary: List all active bonds on market
      operationId: listBonds
      security:
        - BearerAuth: []
      parameters:
        - name: rating
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Bond list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BondListResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Bonds]
      summary: Issue a new bond
      operationId: createBond
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBondRequest'
      responses:
        "200":
          description: Bond created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBondResponse'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/bonds/{bondId}/:
    get:
      tags: [Bonds]
      summary: Get bond detail
      operationId: getBond
      security:
        - BearerAuth: []
      parameters:
        - name: bondId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bond detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBondResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Bond not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v2/companies/me/bonds/owned/:
    get:
      tags: [Bonds]
      summary: Get my owned bonds
      operationId: getOwnedBonds
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Owned bonds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BondListResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v2/companies/me/bonds/sold/:
    get:
      tags: [Bonds]
      summary: Get my issued bonds
      operationId: getSoldBonds
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Issued bonds
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BondListResponse'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        data:
          nullable: true
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object

    # --- Company ---
...
    BuildingDTO:
      type: object
      properties:
        id:
          type: string
        building_id:
          type: integer
        name:
          type: string
        level:
          type: integer
        map_id:
          type: string
        slot_id:
          type: string
        x:
          type: integer
        y:
          type: integer
        robot_count:
          type: integer
        placed:
          type: boolean
        is_retail:
          type: boolean
        produces:
          type: array
          items:
            type: integer
        shelves:
          type: array
          items:
            $ref: '#/components/schemas/ShelfItem'
    BuildingListResponse:
      type: object
      properties:
        buildings:
          type: array
          items:
            $ref: '#/components/schemas/BuildingDTO'
    # --- Building Operations ---
...
    BuyBuildingRequest:
      type: object
      required: [buildingId]
      properties:
        requestId:
          type: string
        buildingId:
          type: string
    BuyBuildingResponse:
      type: object
      properties:
        building:
          $ref: '#/components/schemas/BuildingDTO'
        cost:
          type: number
          format: double
    PlaceBuildingRequest:
      type: object
      required: [buildingId]
      properties:
        buildingId:
          type: string
        mapId:
          type: string
        slotId:
          type: string
        x:
          type: integer
        y:
          type: integer
    PlaceBuildingResponse:
      type: object
      properties:
        building:
          $ref: '#/components/schemas/BuildingDTO'
        status:
          type: string
    MoveBuildingRequest:
      type: object
      required: [buildingId]
      properties:
        buildingId:
          type: string
        mapId:
          type: string
        slotId:
          type: string
        x:
          type: integer
        y:
          type: integer
    MoveBuildingResponse:
      type: object
      properties:
        building:
          $ref: '#/components/schemas/BuildingDTO'
        status:
          type: string
    DemolishBuildingRequest:
      type: object
      required: [buildingId]
      properties:
        buildingId:
          type: string
    DemolishBuildingResponse:
      type: object
      properties:
        refund:
          type: number
          format: double
        status:
          type: string
    UpgradeBuildingResponse:
      type: object
      properties:
        building_id:
          type: string
        old_level:
          type: integer
        new_level:
          type: integer
        cost:
          type: number
          format: double
        output_multiplier:
          type: integer
    ShelfItem:
      type: object
      properties:
        resource_id:
          type: integer
        quantity:
          type: integer
        max_qty:
          type: integer
        price:
          type: number
          format: double
        price_lock:
          type: boolean
        revenue:
          type: number
          format: double
    StockShelfRequest:
      type: object
      required: [building_id, resource_id, quantity]
      properties:
        building_id:
          type: string
        resource_id:
          type: integer
        quantity:
          type: integer
        price:
          type: number
          format: double
    UnstockShelfRequest:
      type: object
      required: [building_id, resource_id, quantity]
      properties:
        building_id:
          type: string
        resource_id:
          type: integer
        quantity:
          type: integer
    SetShelfPriceRequest:
      type: object
      required: [building_id, resource_id, price]
      properties:
        building_id:
          type: string
        resource_id:
          type: integer
        price:
          type: number
          format: double
        lock:
          type: boolean
    ShelfActionResponse:
      type: object
      properties:
        shelf:
          $ref: '#/components/schemas/ShelfItem'
...
    BondDTO:
      type: object
      properties:
        id:
          type: string
        amount:
          type: integer
        interest:
          type: number
        purchased_at:
          type: string
          format: date-time
        missed_payments:
          type: integer
        interestCollected:
          type: number
        ratingWhenPurchased:
          type: string
        dailyInterest:
          type: number
        periodInterest:
          type: number
        issuerCompanyId:
          type: integer
        ownerCompanyId:
          type: integer
        callableAfter:
          type: string
          format: date-time
        restructure_percentage:
          type: number
    BondListResponse:
      type: object
      properties:
        bonds:
          type: array
          items:
            $ref: '#/components/schemas/BondDTO'
    GetBondResponse:
      type: object
      properties:
        bond:
          $ref: '#/components/schemas/BondDTO'
    CreateBondRequest:
      type: object
      required: [amount, interest]
      properties:
        amount:
          type: integer
          minimum: 1
        interest:
          type: number
          description: Interest percent (e.g. 1.5 means 1.5%)
    CreateBondResponse:
      type: object
      properties:
        bond:
          $ref: '#/components/schemas/BondDTO'